# design/openapi.yaml
openapi: 3.0.0
info:
  title: Soroco House — Restaurant Order Management API
  version: 1.0.0
  description: |
    Backend API for the Soroco House QR-menu ordering system.

    Handles:
    - Customer flows: menu browsing, order placement with sandbox/mock payment,
      bill email.
    - Staff flows: login/signup (JWT with role), kitchen order board (FIFO),
      live updates over WebSocket.
    - Admin flows: menu CRUD, image upload, employee management, order history
      + CSV export.

    Conventions:
    - All JSON routes are under /api/v1 (this server URL). Exceptions: GET /health
      (raw health object) and the static /images mount.
    - Every JSON success uses the envelope {"success": true, "data": ...};
      every failure uses {"success": false, "message", "code", "detail"}.
    - Enums are stored as plain VARCHAR values (employee/admin, phonepay/razorpay,
      success/failed/cancelled, in_queue/preparing/prepared/delivered,
      ordered/delivered, standard/small/large).
    - Menu create/delete and image upload are ADMIN only; menu PATCH and kitchen
      routes allow employee + admin.
    - Prices are always recomputed server-side at POST /payments (5% tax).
    - Rate limits: auth 5/min, upload 5/min, payments 20/min.
    - Real-time events stream over WebSockets (documented in design/api-docs.md):
      ws /api/v1/ws/menu (public, frame {"type":"menu_updated"}) and
      ws /api/v1/ws/orders?token=<JWT> (frame {"type":"orders_updated","data":[...]}).
...
servers:
...
paths:
...
  /menu:
    get:
      summary: List menu grouped by category
      description: |
        Public. Returns the full menu grouped by category. Unavailable items
        (is_available=false) are still returned — greyed out client-side for
        customers, visible to staff so they can re-enable.
      operationId: listMenu
      tags: [Menu]
      security: []
      responses:
        '200':
          description: Menu groups
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
              examples:
                ok:
                  value:
                    success: true
                    data:
                      - category: "Hot Luxury Teas"
                        items:
                          - menu_uuid: "m1111111-1111-1111-1111-111111111111"
                            menu_id: 1
                            category: "Hot Luxury Teas"
                            item_name: "Darjeeling First Flush"
                            item_description: "Light, floral & muscatel."
                            standard_price: 280.0
                            small_price: null
                            large_price: null
                            image_url: "https://images.unsplash.com/...?w=640&q=85"
                            is_available: true
                      - category: "Cold Brew"
                        items:
                          - menu_uuid: "m2222222-2222-2222-2222-222222222222"
                            menu_id: 9
                            category: "Cold Brew"
                            item_name: "Signature Cold Brew"
                            item_description: "16-hour slow-steeped cold brew."
                            standard_price: 220.0
                            small_price: 220.0
                            large_price: 320.0
                            image_url: null
                            is_available: true

    post:
      summary: Add a new menu item
      description: ADMIN only. Broadcasts menu_updated to /ws/menu.
      operationId: addMenuItem
      tags: [Menu]
      security:
        - bearer_auth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MenuItemCreate'
            examples:
              valid:
                value:
                  name: "Mango Frappe"
                  category: "Frappe"
                  description: "Seasonal"
                  standard_price: 300.0
                  small_price: null
                  large_price: 360.0
      responses:
        '200':
          description: Menu item created (200 — not 201)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
              examples:
                created:
                  value:
                    success: true
                    data:
                      menu_uuid: "m3333333-3333-3333-3333-333333333333"
                      menu_id: 25
                      category: "Frappe"
                      item_name: "Mango Frappe"
                      item_description: "Seasonal"
                      standard_price: 300.0
                      small_price: null
                      large_price: 360.0
                      image_url: null
                      is_available: true
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
        '409':
          $ref: '#/components/responses/conflict'
        '422':
          $ref: '#/components/responses/validation_error'
  /menu/{menu_uuid}:
    patch:
      summary: Update a menu item (availability, prices, text)
      description: Employee + admin. All fields optional. Broadcasts to /ws/menu.
      operationId: updateMenuItem
      tags: [Menu]
      security:
        - bearer_auth: []
      parameters:
        - name: menu_uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
            example: "m1111111-1111-1111-1111-111111111111"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MenuItemPatch'
            examples:
              toggle:
                value: { is_available: false }
              prices:
                value: { small_price: 240.0, large_price: 340.0 }
      responses:
        '200':
          description: Menu item updated (full item)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
              examples:
                updated:
                  value:
                    success: true
                    data:
                      menu_uuid: "m1111111-1111-1111-1111-111111111111"
                      menu_id: 1
                      category: "Hot Luxury Teas"
                      item_name: "Darjeeling First Flush"
                      item_description: "Light, floral & muscatel."
                      standard_price: 280.0
                      small_price: null
                      large_price: null
                      image_url: null
                      is_available: false
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
        '404':
          $ref: '#/components/responses/not_found'
        '422':
          $ref: '#/components/responses/validation_error'

    delete:
      summary: Remove a menu item
      description: ADMIN only. Broadcasts to /ws/menu.
      operationId: deleteMenuItem
      tags: [Menu]
      security:
        - bearer_auth: []
      parameters:
        - name: menu_uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
            example: "m1111111-1111-1111-1111-111111111111"
      responses:
        '200':
          description: Menu item deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessageEnvelope'
              examples:
                deleted:
                  value:
                    success: true
                    data:
                      message: "Menu item deleted"
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
        '404':
          $ref: '#/components/responses/not_found'

  # ─── Uploads ───
...
  /orders/kitchen:
    get:
      summary: List kitchen orders FIFO
      description: |
        Employee + admin. Returns all orders, oldest first. Each order carries
        its persisted kitchen/order status so the board restores state after a
        refresh.
      operationId: listKitchenOrders
      tags: [Orders]
      security:
        - bearer_auth: []
      responses:
        '200':
          description: Array of serialized orders
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
              examples:
                ok:
                  value:
                    success: true
                    data:
                      - order_uuid: "o9999999-9999-9999-9999-999999999999"
                        order_id: 5
                        order_number: 12
                        table_name: "A1"
                        customer_name: "Alex"
                        phone_number: "+919876543210"
                        payment_method: "razorpay"
                        payment_status: "success"
                        payment_transaction_id: "pay_PkF21XS9mup1b9"
                        kitchen_status: "in_queue"
                        order_status: "ordered"
                        total_price: 483.0
                        subtotal: 460.0
                        tax: 23.0
                        created_at: "2026-09-20T10:36:12.123456+00:00"
                        items:
                          - menu_uuid: "m2222222-2222-2222-2222-222222222222"
                            item_name: "Signature Cold Brew"
                            selected_size: "large"
                            quantity: 2
                            unit_price: 220.0
                            line_total: 440.0
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
  /orders/{order_uuid}:
    patch:
      summary: Update order status (kitchen board)
      description: |
        Employee + admin. Updates kitchen_status
        (in_queue | preparing | prepared | delivered) and/or order_status
        (ordered | delivered). Enum-only validation — no enforced transition
        order. Broadcasts the fresh kitchen list to /ws/orders.
      operationId: updateOrderStatus
      tags: [Orders]
      security:
        - bearer_auth: []
      parameters:
        - name: order_uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
            example: "o9999999-9999-9999-9999-999999999999"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderPatch'
            examples:
              preparing:
                value: { kitchen_status: "preparing" }
              delivered:
                value: { order_status: "delivered" }
      responses:
        '200':
          description: Full updated order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
              examples:
                updated:
                  value:
                    success: true
                    data:
                      order_uuid: "o9999999-9999-9999-9999-999999999999"
                      order_id: 5
                      order_number: 12
                      table_name: "A1"
                      customer_name: "Alex"
                      phone_number: "+919876543210"
                      payment_method: "razorpay"
                      payment_status: "success"
                      payment_transaction_id: "pay_PkF21XS9mup1b9"
                      kitchen_status: "preparing"
                      order_status: "ordered"
                      total_price: 483.0
                      subtotal: 460.0
                      tax: 23.0
                      created_at: "2026-09-20T10:36:12.123456+00:00"
                      items: []
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
        '404':
          $ref: '#/components/responses/not_found'
        '422':
          $ref: '#/components/responses/validation_error'

  # ─── Admin: Employees ───
...
components:
  securitySchemes:
    bearer_auth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT issued by POST /auth/login or POST /auth/signup.
  responses:
    unauthorized:
      description: Missing, invalid or expired token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            bad_token:
              value:
                success: false
                message: "Not authenticated"
                code: "UNAUTHORIZED"
                detail: null
    forbidden:
      description: Role insufficient for this route
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            role:
              value:
                success: false
                message: "You don't have permission to perform this action"
                code: "FORBIDDEN"
                detail: null
    not_found:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            missing:
              value:
                success: false
                message: "Menu item not found"
                code: "NOT_FOUND"
                detail: null
    conflict:
      description: Duplicate / forbidden-state conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            dup:
              value:
                success: false
                message: "An account already exists for that email"
                code: "CONFLICT"
                detail: null
    validation_error:
      description: Pydantic or business-rule validation failure
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            invalid:
              value:
                success: false
                message: "phone_number: Enter a valid phone number (10 digits)"
                code: "VALIDATION_ERROR"
                detail: null
...
  schemas:
    SuccessEnvelope:
      type: object
      description: Standard success envelope — data carries the payload.
      required: [success, data]
      properties:
        success:
          type: boolean
          const: true
          example: true
        data:
          nullable: true
          description: Payload (object, array, string or number depending on endpoint).
    SuccessMessageEnvelope:
      type: object
      description: Success envelope for message-only responses (ok_message helper).
      required: [success, data]
      properties:
        success:
          type: boolean
          const: true
          example: true
        data:
          type: object
          required: [message]
          properties:
            message:
              type: string
              example: "Employee removed"
    ErrorEnvelope:
      type: object
      description: Standard error envelope for every non-2xx JSON response.
      required: [success, message, code, detail]
      properties:
        success:
          type: boolean
          const: false
          example: false
        message:
          type: string
          example: "Menu item not found"
        code:
          type: string
          nullable: true
          enum: [UNAUTHORIZED, INVALID_CREDENTIALS, FORBIDDEN, SIGNUP_PENDING, NOT_FOUND, CONFLICT, VALIDATION_ERROR, RATE_LIMITED, INTERNAL_ERROR, FORGED_CALLBACK]
          example: "NOT_FOUND"
        detail:
          type: string
          nullable: true
          example: null
...
    MenuItemCreate:
      type: object
      description: |
        Request field names are name/description (responses use
        item_name/item_description). A convenience `price` alias maps to
        standard_price. At least one price is required.
      required: [name, category]
      properties:
        name:
          type: string
        category:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: number
          nullable: true
          description: Alias for standard_price
        standard_price:
          type: number
          nullable: true
        small_price:
          type: number
          nullable: true
        large_price:
          type: number
          nullable: true
        image_url:
          type: string
          nullable: true
    MenuItemPatch:
      type: object
      description: At least one field required.
      properties:
        name:
          type: string
        category:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: number
          nullable: true
          description: Alias for standard_price
        standard_price:
          type: number
          nullable: true
        small_price:
          type: number
          nullable: true
        large_price:
          type: number
          nullable: true
        image_url:
          type: string
          nullable: true
        is_available:
          type: boolean
...
    OrderPatch:
      type: object
      properties:
        kitchen_status:
          type: string
          enum: [in_queue, preparing, prepared, delivered]
        order_status:
          type: string
          enum: [ordered, delivered]
...